// resources/js/api/generated/models/operationLogInfo.ts
/**
 * Generated by orval v8.0.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */
import type { DeletedAt } from './deletedAt';
import type { Server } from './server';
import type { User } from './user';

export interface OperationLogInfo {
  command: string;
  created_at: string;
  deleted_at?: DeletedAt;
  /** @nullable */
  duration_ms?: number | null;
  /** @nullable */
  end_time?: string | null;
  /** @nullable */
  exit_code?: number | null;
  formatted_date: string;
  /** @minimum 0 */
  id: number;
  is_incomplete: boolean;
  /** @nullable */
  last_message_at?: string | null;
  message_count: number;
  operation_id: string;
  options?: string;
  /** @nullable */
  partial_duration_ms?: number | null;
  /** @nullable */
  queued_at?: string | null;
  server: Server;
  /** @minimum 0 */
  server_id: number;
  server_name: string;
  services?: string;
  stack_name: string;
  start_time?: string;
  status?: string;
  /** @nullable */
  success?: boolean | null;
  summary?: string;
  trigger_source: string;
  updated_at: string;
  user: User;
  /** @minimum 0 */
  user_id: number;
  user_name: string;
}